fix: icon 컴포넌트를 파일별로 분리
