+++ src/basic/components/icon/LargeBagIcon.tsx
@@ -1,0 +1,16 @@
+// 장바구니 아이콘 (large)
+export const LargeBagIcon = () => (
+  <svg
+    className="w-16 h-16 text-gray-300 mx-auto mb-4"
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={1}
+      d="M16 11V7a4 4 0 00-8 0v4M5 9h14l1 12H4L5 9z"
+    />
+  </svg>
+);
