+++ src/basic/components/icon/PlusIcon.tsx
@@ -1,0 +1,16 @@
+// + 버튼 아이콘
+export const PlusIcon = () => (
+  <svg
+    className="w-8 h-8"
+    fill="none"
+    stroke="currentColor"
+    viewBox="0 0 24 24"
+  >
+    <path
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={2}
+      d="M12 4v16m8-8H4"
+    />
+  </svg>
+);
