+++ src/basic/components/icon/index.ts
@@ -1,0 +1,7 @@
+export * from "./CartIcon";
+export * from "./CloseIcon";
+export * from "./ImageIcon";
+export * from "./LargeBagIcon";
+export * from "./PlusIcon";
+export * from "./SmallBagIcon";
+export * from "./TrashIcon";
